@@ -13,7 +13,10 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.0|^5.0"
     },
     "autoload": {
         "psr-4": {
@@ -22,7 +25,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Endroid\\Bundle\\IntakeBundle\\": "tests/"
+            "Endroid\\Bundle\\IntakeBundle\\Tests\\": "tests/"
         }
     },
     "config": {
